% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_values.R
\docType{data}
\name{M_values}
\alias{M_values}
\title{M-values for 1000 DNA methylation sites}
\format{
A data frame with 1000 rows (corresponding to DNA methyaltion sites)
and 128 columns (corresponding to persons in ped with typed = 1).
}
\source{
Simulated
}
\usage{
M_values
}
\description{
A dataset containing simulated M-values for 1000 DNA methylation sites
on selected people from the ped dataset (those with typed = 1).
}
\keyword{datasets}
